+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroExtensionRegistrarConfigurator.kt
@@ -84,6 +84,9 @@ class MetroExtensionRegistrarConfigurator(testServices: TestServices) :
         optionalDependencyBehavior =
           module.directives.singleOrZeroValue(MetroDirectives.OPTIONAL_DEPENDENCY_BEHAVIOR)
             ?: optionDefaults.optionalDependencyBehavior,
+        interopAnnotationsNamedArgSeverity =
+          module.directives.singleOrZeroValue(MetroDirectives.INTEROP_ANNOTATIONS_NAMED_ARG_SEVERITY)
+            ?: optionDefaults.interopAnnotationsNamedArgSeverity,
         maxIrErrorsCount =
           module.directives.singleOrZeroValue(MetroDirectives.MAX_IR_ERRORS_COUNT) ?: 20,
         enableDaggerAnvilInterop = MetroDirectives.WITH_ANVIL in module.directives,